backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-multipart
PyJWT[crypto]
passlib[bcrypt]
supabase==2.3.0
python-dotenv
numpy
numpy-financial

# Security additions
slowapi
starlette
pydantic[email]
